pyproject.toml
[project]
name = "notemine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"

requires-python = ">=3.13"
dependencies = [
    "matplotlib>=3.7.0",
    "plotly>=5.0.0",
    "numpy>=2.3.4",
    "pandas>=2.0.0",
    "python-dotenv>=1.0.0",
    "torch",
    "torchvision",
    "torchaudio",
    "pyyaml>=6.0.3",
    "openai>=2.31.0",
    "anthropic>=0.93.0",
    "python-frontmatter>=1.1.0",
]

[tool.uv.sources]
torch = { index = "pytorch-cu124" }
torchvision = { index = "pytorch-cu124" }
torchaudio = { index = "pytorch-cu124" }

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true


[dependency-groups]
notebook = [
    "ipykernel>=6.0.0",
    "notebook>=7.0.0",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
